Get the OPSCTL_GITHUB_TOKEN env var from the actual environment if --token is not passed to the generate command
This is most likely set locally anyway and --token is not needed with the split setup.

When config-controller runs in controller mode, the vault client is created in a different way
so the same token is not needed there. In CLI mode it uses opsctl which requires the token to be set.